requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
whitenoise==5.2.0
